split component and hook
