+++ packages/app/src/cli/services/function/ui/components/Replay.tsx
@@ -1,19 +1,13 @@
+import {setupExtensionWatcherForReplay} from './hooks/extension-watcher.js'
 import {FunctionRunData} from '../../replay.js'
-import {setupExtensionWatcher} from '../../../dev/extension/bundler.js'
 import {ExtensionInstance} from '../../../../models/extensions/extension-instance.js'
 import {FunctionConfigType} from '../../../../models/extensions/specifications/function.js'
 import {AppInterface} from '../../../../models/app/app.js'
 import {prettyPrintJsonIfPossible} from '../../../app-logs/utils.js'
-import {exec} from '@shopify/cli-kit/node/system'
 import figures from '@shopify/cli-kit/node/figures'
 import {AbortController} from '@shopify/cli-kit/node/abort'
-import React, {FunctionComponent, useEffect, useMemo, useRef, useState} from 'react'
-import {Box, Text, Static, useInput, useStdin} from '@shopify/cli-kit/node/ink'
-import {handleCtrlC} from '@shopify/cli-kit/node/ui'
-import {useAbortSignal} from '@shopify/cli-kit/node/ui/hooks'
-import {isUnitTest} from '@shopify/cli-kit/node/context/local'
-import {treeKill} from '@shopify/cli-kit/node/tree-kill'
-import {Writable} from 'stream'
+import React, {FunctionComponent} from 'react'
+import {Box, Text, Static} from '@shopify/cli-kit/node/ink'
 
 export interface ReplayProps {
   selectedRun: FunctionRunData
@@ -41,122 +35,12 @@ interface SystemMessage {
 type ReplayLog = FunctionRun | SystemMessage
 
 const Replay: FunctionComponent<ReplayProps> = ({selectedRun, abortController, app, extension}) => {
-
-  const functionRunFromSelectedRun = {
-    type: 'functionRun',
-    input: selectedRun.payload.input,
-    output: selectedRun.payload.output,
-    logs: selectedRun.payload.logs,
-    name: selectedRun.source,
-    size: 0,
-    memory_usage: 0,
-    instructions: selectedRun.payload.fuelConsumed,
-  } as FunctionRun
-
-  const [logs, setLogs] = useState<ReplayLog[]>([])
-  const [recentFunctionRuns, setRecentFunctionRuns] = useState<[FunctionRun, FunctionRun]>([
-    functionRunFromSelectedRun,
-    functionRunFromSelectedRun,
-  ])
-
-  const [error, setError] = useState<string | undefined>(undefined)
-
-  const {input, export: runExport} = selectedRun.payload
-
-  const {isRawModeSupported: canUseShortcuts} = useStdin()
-  const pollingInterval = useRef<NodeJS.Timeout>()
-
-  const [statusMessage, setStatusMessage] = useState(`Watching for changes to ${selectedRun.source}...`)
-
-  useEffect(() => {
-    // run the selectedRun once
-    const initialReplay =  async () => {
-      setStatusMessage("Replaying log with local function...")
-      const functionRun = await runFunctionRunnerWithLogInput(extension, JSON.stringify(input), runExport)
-      setRecentFunctionRuns((recentFunctionRuns) => {
-        return [functionRun, recentFunctionRuns[0]]
-      })
-      setStatusMessage(`Watching for changes to ${selectedRun.source}...`)
-      setLogs((logs) => [...logs, functionRun])
-    }
-
-
-    const startWatchingFunction = async () => {
-      const customStdout = new Writable({
-        write(chunk, _enconding, next) {
-          setLogs((logs) => [...logs, {type: 'systemMessage', message: chunk.toString()}])
-          next()
-        },
-      })
-
-      await setupExtensionWatcher({
-        extension,
-        app,
-        stdout: customStdout, // TODO
-        stderr: customStdout, // TODO
-        onChange: async () => {
-          setStatusMessage("Re-running with latest changes...")
-          const functionRun = await runFunctionRunnerWithLogInput(extension, JSON.stringify(input), runExport)
-
-          setRecentFunctionRuns((recentFunctionRuns) => {
-            return [functionRun, recentFunctionRuns[0]]
-          })
-
-          setStatusMessage(`Watching for changes to ${selectedRun.source}...`)
-          setLogs((logs) => [...logs, functionRun])
-        },
-        onReloadAndBuildError: async (error) => {
-          // TODO: handle error
-        },
-        signal: abortController.signal,
-      })
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    initialReplay()
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    startWatchingFunction()
-
-    // TODO: return a way to clean up watcher
-  }, [input, runExport, app, extension])
-
-  const {isAborted} = useAbortSignal(abortController.signal, async (err) => {
-    if (err) {
-      setStatusMessage('Shutting down replay watcher because of an error ...')
-    } else {
-      setStatusMessage('Shutting down replay watcher ...')
-      setTimeout(() => {
-        if (isUnitTest()) return
-        treeKill(process.pid, 'SIGINT', false, () => {
-          process.exit(0)
-        })
-      }, 2000)
-    }
-    clearInterval(pollingInterval.current)
+  const {logs, isAborted, canUseShortcuts, statusMessage, recentFunctionRuns} = setupExtensionWatcherForReplay({
+    selectedRun,
+    abortController,
+    app,
+    extension,
   })
-
-  useInput(
-    (input, key) => {
-      handleCtrlC(input, key, () => abortController.abort())
-
-      const onInput = async () => {
-        try {
-          setError('')
-
-          if (input === 'q') {
-            abortController.abort()
-          }
-          // eslint-disable-next-line no-catch-all/no-catch-all
-        } catch (_) {
-          setError('Failed to handle your input.')
-        }
-      }
-
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      onInput()
-    },
-    {isActive: Boolean(canUseShortcuts)},
-  )
 
   return (
     <>
@@ -173,32 +57,32 @@ const Replay: FunctionComponent<ReplayProps> = ({selectedRun, abortController, a
       {/* Bottom Bar */}
       {/* eslint-disable-next-line no-negated-condition */}
       {!isAborted ? (
-      <Box
-        marginY={1}
-        paddingTop={1}
-        flexDirection="column"
-        flexGrow={1}
-        borderStyle="single"
-        borderBottom={false}
-        borderLeft={false}
-        borderRight={false}
-        borderTop
-      >
-        {canUseShortcuts ? (
-          <Box flexDirection="column">
-            <Box flexDirection="row">
+        <Box
+          marginY={1}
+          paddingTop={1}
+          flexDirection="column"
+          flexGrow={1}
+          borderStyle="single"
+          borderBottom={false}
+          borderLeft={false}
+          borderRight={false}
+          borderTop
+        >
+          {canUseShortcuts ? (
+            <Box flexDirection="column">
+              <Box flexDirection="row">
+                <Text>
+                  {figures.pointerSmall} {statusMessage}
+                </Text>
+              </Box>
+              <StatsDisplay recentFunctionRuns={recentFunctionRuns} />
               <Text>
-                {figures.pointerSmall} {statusMessage}
+                {figures.pointerSmall} Press <Text bold>q</Text> {figures.lineVertical} quit
               </Text>
             </Box>
-            <StatsDisplay recentFunctionRuns={recentFunctionRuns}/>
-            <Text>
-              {figures.pointerSmall} Press <Text bold>q</Text> {figures.lineVertical} quit
-            </Text>
-          </Box>
-        ) : null}
-      </Box>
-      ) : null }
+          ) : null}
+        </Box>
+      ) : null}
     </>
   )
 }
@@ -265,10 +149,10 @@ function ReplayLog({log}: {log: ReplayLog}) {
   if (log.type === 'functionRun') {
     return (
       <Box flexDirection="column">
-        <InputDisplay input={log.input}/>
-        <LogDisplay logs={log.logs}/>
-        <OutputDisplay output={log.output}/>
-        <BenchmarkDisplay functionRun={log}/>
+        <InputDisplay input={log.input} />
+        <LogDisplay logs={log.logs} />
+        <OutputDisplay output={log.output} />
+        <BenchmarkDisplay functionRun={log} />
       </Box>
     )
   }
@@ -281,27 +165,3 @@ function ReplayLog({log}: {log: ReplayLog}) {
 }
 
 export {Replay}
-
-async function runFunctionRunnerWithLogInput(
-  fun: ExtensionInstance<FunctionConfigType>,
-  input: string,
-  exportName: string,
-): Promise<FunctionRun> {
-  let functionRunnerOutput = ''
-  const customStdout = new Writable({
-    write(chunk, _encoding, next) {
-      functionRunnerOutput += chunk
-      next()
-    },
-  })
-
-  await exec('npm', ['exec', '--', 'function-runner', '--json', '-f', fun.outputPath, '--export', exportName], {
-    cwd: fun.directory,
-    input,
-    stdout: customStdout,
-    stderr: 'inherit',
-  })
-
-  const result = JSON.parse(functionRunnerOutput)
-  return {...result, type: 'functionRun'}
-}
